+++ backend/canisters/local_user_index/api/src/updates/c2c_create_multi_user_canister.rs
@@ -1,0 +1,12 @@
+use oc_error_codes::OCError;
+use serde::{Deserialize, Serialize};
+use types::CanisterId;
+
+#[derive(Serialize, Deserialize, Debug)]
+pub struct Args {}
+
+#[derive(Serialize, Deserialize, Debug)]
+pub enum Response {
+    Success(CanisterId),
+    Error(OCError),
+}