+++ backend/integration_tests/src/wasms.rs
@@ -17,6 +17,7 @@ lazy_static! {
     pub static ref ICRC_LEDGER: CanisterWasm = get_canister_wasm("icrc_ledger");
     pub static ref IDENTITY: CanisterWasm = get_canister_wasm("identity");
     pub static ref LOCAL_USER_INDEX: CanisterWasm = get_canister_wasm("local_user_index");
+    pub static ref MULTI_USER: CanisterWasm = get_canister_wasm("multi_user");
     pub static ref NOTIFICATIONS_INDEX: CanisterWasm = get_canister_wasm("notifications_index");
     pub static ref ONLINE_USERS: CanisterWasm = get_canister_wasm("online_users");
     pub static ref OPENCHAT_INSTALLER: CanisterWasm = get_canister_wasm("openchat_installer");